[ADD] BoardDetailViewModelFactory
- 차후 공지 관련 ViewModel에서도 사용될 예정이기 때문에, 갈아 끼워 적용할 수 있도록 팩토리 패턴을 구현하였습니다.
